add join course with input
