+++ packages/service-mobile-app/src/pages/courses/Routes.tsx
@@ -1,6 +1,7 @@
 import { Route, Routes } from 'react-router-native';
 import CourseDashboard from 'src/pages/courses/dashboard/CourseDashboard/CourseDashboard';
 import SectionDashboard from 'src/pages/courses/dashboard/SectionDashboard/SectionDashboard';
+import FindCourse from 'src/pages/courses/FindCourse/FindCourse';
 import JoinCourse from 'src/pages/courses/JoinCourse/JoinCourse';
 
 import Lesson from './dashboard/lesson';
@@ -8,7 +9,8 @@ import Lesson from './dashboard/lesson';
 function CourseRouter() {
   return (
     <Routes>
-      <Route path="join/:id" element={<JoinCourse />} />
+      <Route path=":id/join" element={<JoinCourse />} />
+      <Route path="find" element={<FindCourse />} />
 
       <Route path=":id" element={<CourseDashboard />} />
       <Route path=":id/section/:sectionId" element={<SectionDashboard />} />
